% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{build_tcga_mat}
\alias{build_tcga_mat}
\title{Create an annotated matrix of quantifications for a specified TCGA tumor site and assay}
\usage{
build_tcga_mat(site, brief_assay = "rsem_genes_norm",
  assay = assay_cands[brief_assay], localize_dry = FALSE,
  suffix = "_terra_txt", choice = NULL)
}
\arguments{
\item{site}{character(1) anatomic site of tumor}

\item{brief_assay}{character(1) element of names(gdc_assay_map())}

\item{assay}{character(1) full name of assay in GDC nomenclature}

\item{localize_dry}{logical(1) set value of parameter `dry` for AnVIL::localize call; default FALSE implies that data will be retrieved}

\item{suffix}{character(1) used to annotate folder name where text files are localized}

\item{choice}{numeric(1) function fails if AnVIL has multiple versions
of assay quantifications in place, and will report in order the names
of versions present; set `choice` to the position in the
failure report of the version to be used}
}
\value{
a matrix with rownames determined by column 1 of
the first discovered text file (skipping first line)
and colnames determined by the identifier present in the
first line of each text file
}
\description{
Create an annotated matrix of quantifications for a specified TCGA tumor site and assay
}
\note{
If `localize_dry` is TRUE, NULL is returned.
}
\examples{
if (interactive()) { # will egress
 curd = getwd()
 td = tempdir()
 setwd(td)
 tst = build_tcga_mat("acc", "rppa", suffix="accrppatst")
 tst[1:4,1:4]
 fis = dir()
 nf = grep("accrppatst", fis, value=TRUE)
 if (length(nf)==1) unlink(nf, recursive=TRUE)
 setwd(curd)
}
}
