% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{gdc_assay_map}
\alias{gdc_assay_map}
\title{map assay names used in GDC to shorter (improvised) names}
\usage{
gdc_assay_map()
}
\value{
character vector
}
\description{
map assay names used in GDC to shorter (improvised) names
}
\examples{
gdmap = gdc_assay_map()
summary(nchar(gdmap))
head(gdmap)
}
