% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{localize_tumor_site}
\alias{localize_tumor_site}
\title{retrieve GDC text files from AnVIL buckets}
\usage{
localize_tumor_site(site, brief_assay = "rsem_genes_norm",
  assay = assay_cands[brief_assay], suffix = "_terra_txt",
  localize_dry = FALSE, choice = NULL)
}
\arguments{
\item{site}{character(1) anatomic site of tumor}

\item{brief_assay}{character(1) element of names(gdc_assay_map())}

\item{assay}{character(1) full name of assay in GDC nomenclature}

\item{suffix}{character(1) used to annotate folder name where text files are localized}

\item{localize_dry}{logical(1) set value of parameter `dry` for AnVIL::localize call; default FALSE implies that data will be retrieved}

\item{choice}{numeric(1) function fails if AnVIL has multiple versions
of assay quantifications in place, and will report in order the names
of versions present; set `choice` to the position in the
failure report of the version to be used}
}
\value{
invisibly, the name of the folder where text files
were localized
}
\description{
retrieve GDC text files from AnVIL buckets
}
\note{
When successful, this function populates a new folder
with text files drawn from GCP buckets managed in the AnVIL
project.  A tempdir() could be used, this may be default in future,
so this function may be redesigned in the near future.
}
\examples{
if (interactive()) { # will egress
curd = getwd()
td = tempdir()
setwd(td)
nn = localize_tumor_site(site="acc", brief_assay="rppa",
 localize_dry = TRUE)
dir(nn)
unlink(nn, recursive=TRUE)
setwd(curd)
}
}
