% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{tcga_sites}
\alias{tcga_sites}
\title{enumerate all TCGA tumor sites (anatomic site codes)}
\usage{
tcga_sites()
}
\value{
character vector
}
\description{
enumerate all TCGA tumor sites (anatomic site codes)
}
\examples{
head(tcga_sites())
}
